adds access restrictions based on resource ownership

+++ app/graphql/mutations/delete_encounter.rb
@@ -9,6 +9,7 @@ module Mutations
       authorize_user
 
       encounter = Encounter.find(id)
+      return if context[:current_user].id != encounter.user_id
       raise GraphQL::ExecutionError, encounter.errors.full_messages.join(', ') unless encounter.destroy
 
       { id: }
